refactor(spk-440): extract resolveColorPalette and apply cascade to SQL charts
Lifts the chart → space → dashboard → project → org cascade out of
SavedChartModel into a shared resolveColorPalette helper in
organizationColorPalettes.ts so SQL charts can reuse the same logic.
SavedSqlService.getSavedSqlChart now returns resolvedColorPalette on
the SqlChart payload (chart-uuid is omitted since SQL charts inherit
only), and ContentPanel reads it for the SQL Runner editor view.
Tests for the cascade move to organizationColorPalettes.test.ts; the
two existing palette tests in SavedChartModel.test.ts are deleted.
